Move searchreplace to utils/
